Fix poller not receiving events on Windows
The `polling` crate uses the IOCP mechanism on Windows to support async-IO. Unfortunately, WinPTY doesn't support IOCP, so `poller.add()` doesn't work.

`alacritty_terminal` worked around the issue by creating a worker thread that constantly pull data from the PTY and post events to the poller manually, effectively using the poller as a simple event queue.

A limitation to this approach is that we have to attempt a read before calling `poller.wait()`, otherwise no event will be delivered.